Apply scope `if-block` only if a line starts with `if`.
(Leading whitespace is allowed, though.)

This fixes “runaway” scopes for lines like `dd if=/dev/sda1…`.

Runaway scopes are especially a problem if the grammar is embedded in other grammars. For example, in the markdown-redcarpet bundle, if there is a code block triggering a runaway scope, the block’s grammar will extend until EOF, completely messing up the syntax highlight after the end of the code block.