Custom word boundary matchers
The regexp engine and the shell grammar don't agree on what comprises
a “word boundary”. This led to problems where a command or path
ending in a shell keyword would be interpreted as that keyword.
This change replaces `\b` word boundary matchers with lookahead/-behind
matchers on whitespace, line breaks, and command separators (";" and "&").

Closes #22.