Require terminator before function definition
The current grammar assumes it found a function whenever it encounters the keyword `function` followed by any word, which is too aggressive and can lead to false positives.
This change requires that the function definition is preceded by a terminator (`^|;|&`) and (optional) whitespace.

This approach can still produce false positives (it doesn’t recognise when the line break was escaped, for example) but should work well enough.
It also now includes the preceding whitespace in the `meta.function.shell` scope (due to the limitations on variable-width look-behinds), which I don’t like but also don’t see a way around.

We might want to consider doing something similar for loop constructs.

Closes #36