Improve reliability of tab creation (even more)
Wait until the number of tabs in the terminal application has changed before executing the shell command. Abort after 2 seconds, if no new tab seems to have appeared.

This fixes a race condition where the new tab was opened but the shell command was still run in the previous open tab.

It does not fix the problem with the user not releasing the modifier keys used to initiate the command, though (see https://github.com/textmate/shellscript.tmbundle/commit/32af606a9e4e56707d1be4569947a516aa20bc64). But if this happens, we now at least do not try to run the shell command in an existing tab.