Fix “Run in Terminal” for macOS High Sierra
It seems that in High Sierra, `count tabs` for the Terminal app always returns 1 (probably a bug?). So this can no longer be used for checking if the new tab is ready.

Instead we now wait until the content of the visible tab has changed, which seems to work fine